+++ spring-messaging/src/test/java/org/springframework/messaging/handler/invocation/reactive/MethodMessageHandlerTests.java
@@ -24,7 +24,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Consumer;
-import java.util.function.Predicate;
 
 import org.junit.Test;
 import org.reactivestreams.Publisher;
@@ -203,6 +202,10 @@ public class MethodMessageHandlerTests {
 		private PathMatcher pathMatcher = new AntPathMatcher();
 
 
+		public TestMethodMessageHandler() {
+			setHandlerPredicate(handlerType -> handlerType.getName().endsWith("Controller"));
+		}
+
 		@Override
 		protected List<? extends HandlerMethodArgumentResolver> initArgumentResolvers() {
 			return Collections.emptyList();
@@ -211,11 +214,6 @@ public class MethodMessageHandlerTests {
 		@Override
 		protected List<? extends HandlerMethodReturnValueHandler> initReturnValueHandlers() {
 			return Collections.singletonList(this.returnValueHandler);
-		}
-
-		@Override
-		protected Predicate<Class<?>> initHandlerPredicate() {
-			return handlerType -> handlerType.getName().endsWith("Controller");
 		}
 
 		@Nullable
